refactor(redis): Move redis keys and serializers packages
- Move key specs and constants into `lib.redis.key`
- Move serializers into `lib.redis.serializer`
- Update related files
- Bump version to 0.4.5
