+++ build.gradle.kts
@@ -10,7 +10,7 @@ plugins {
 
 group = "me.loghub"
 
-version = "0.4.4-SNAPSHOT"
+version = "0.4.5-SNAPSHOT"
 
 java { toolchain { languageVersion = JavaLanguageVersion.of(21) } }
