+++ src/main/kotlin/me/loghub/api/lib/redis/serializer/OAuth2JoinInfoRedisSerializer.kt
@@ -1,4 +1,4 @@
-package me.loghub.api.lib.redis
+package me.loghub.api.lib.redis.serializer
 
 import me.loghub.api.dto.auth.join.OAuth2JoinInfoDTO
 import me.loghub.api.entity.user.User
