+++ src/main/kotlin/me/loghub/api/service/article/ArticleDraftService.kt
@@ -1,9 +1,9 @@
 package me.loghub.api.service.article
 
 import me.loghub.api.constant.message.ResponseMessage
-import me.loghub.api.constant.redis.RedisKeys
 import me.loghub.api.dto.common.UpdateDraftDTO
 import me.loghub.api.entity.user.User
+import me.loghub.api.lib.redis.key.RedisKeys
 import me.loghub.api.repository.article.ArticleRepository
 import me.loghub.api.util.checkPermission
 import org.springframework.data.redis.core.RedisTemplate
