+++ src/main/kotlin/me/loghub/api/service/auth/RefreshService.kt
@@ -2,9 +2,9 @@ package me.loghub.api.service.auth
 
 import jakarta.transaction.Transactional
 import me.loghub.api.constant.message.ResponseMessage
-import me.loghub.api.constant.redis.RedisKeys
 import me.loghub.api.dto.auth.token.TokenDTO
 import me.loghub.api.exception.auth.BadRefreshTokenException
+import me.loghub.api.lib.redis.key.RedisKeys
 import me.loghub.api.repository.user.UserRepository
 import me.loghub.api.service.auth.token.TokenService
 import org.springframework.data.redis.core.RedisTemplate
